+++ src/Executor/PostgresExecutor.php
@@ -21,7 +21,7 @@ namespace DatabaseBackup\Executor;
  */
 class PostgresExecutor extends AbstractExecutor
 {
-    protected const EXPORT_BYNARY = 'pg_dump';
+    protected const EXPORT_BINARY = 'pg_dump';
 
     protected const IMPORT_BINARY = 'pg_restore';
 }