+++ config/bootstrap.php
@@ -15,7 +15,6 @@ declare(strict_types=1);
  */
 
 use Cake\Core\Configure;
-use Symfony\Component\Process\ExecutableFinder;
 
 require_once CAKE . 'functions.php';
 
@@ -45,25 +44,6 @@ $defaults = [
     'DatabaseBackup.sqlite.import' => '{{BINARY}} {{DB_NAME}}',
 ];
 Configure::write(array_filter($defaults, fn (string $key): bool => !Configure::check($key), ARRAY_FILTER_USE_KEY));
-
-/**
- * It automatically discovers executables not already set by the user in the configuration.
- *
- * For `mysql` and `mysqldump` executables, it will first look for `mariadb` and `mariadb-dump` executables.
- * It then normally searches all other possible executables canonically.
- */
-//$ExecutableFinder = new ExecutableFinder();
-//foreach (['mariadb' => 'mysql', 'mariadb-dump' => 'mysqldump'] as $executable => $alias) {
-//    if (!Configure::check('DatabaseBackup.binaries.' . $alias)) {
-//        Configure::write('DatabaseBackup.binaries.' . $alias, $ExecutableFinder->find($executable));
-//    }
-//}
-//$executables = array_merge(['bzip2', 'gzip'], ...array_values(array_map('array_values', DATABASE_BACKUP_EXECUTABLES)));
-//foreach ($executables as $executable) {
-//    if (!Configure::check('DatabaseBackup.binaries.' . $executable)) {
-//        Configure::write('DatabaseBackup.binaries.' . $executable, $ExecutableFinder->find($executable));
-//    }
-//}
 
 //Checks for the target directory
 $target = Configure::read('DatabaseBackup.target');