+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## 2.15 branch
 ### 2.15.0
 * added `AbstractExecutor::findBinary()` method. The `AbstractExecutor::getBinary()` method no longer exists, as it has
-  been completely replaced by the first.
+  been completely replaced by the first;
+* the "Executors" classes now implement the `EXPORT_BINARY` and `IMPORT_BINARY` constants, which are strings or arrays
+  of strings (e.g. in the case of `MysqlExecutor`).
 
 ## 2.14 branch
 ### 2.14.1