+++ CHANGELOG.md
@@ -7,7 +7,8 @@
   `getImportBinary()` methods, which should return the names of the binaries (as a string or array of strings) related
   to the respective driver;
 * the "Executors" classes now implement the `EXPORT_BINARY` and `IMPORT_BINARY` constants, which are strings or arrays
-  of strings (e.g. in the case of `MysqlExecutor`).
+  of strings (e.g. in the case of `MysqlExecutor`);
+* `BackupTrait` had been deprecated and has now been removed.
 
 ## 2.14 branch
 ### 2.14.1