+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   to the respective driver;
 * the "Executors" classes now implement the `EXPORT_BINARY` and `IMPORT_BINARY` constants, which are strings or arrays
   of strings (e.g. in the case of `MysqlExecutor`);
+* the `AbstractBackupUtility::__get()` magic method, which allowed to read the properties of the "Executor" classes, had
+  been deprecated and has now been removed;
 * `BackupTrait` had been deprecated and has now been removed;
 * aliases for old `Driver` classes have been removed.